gui.py: Split into modules [to be continued]
