+++ Orange/widgets/gui/__init__.py
@@ -1,0 +1,26 @@
+# modules define __all__
+# pylint: disable=wildcard-import
+
+from .boxes import *
+from .buttons import *
+from .checkbox import *
+from .combobox import *
+from .labels import *
+from .lineedit import *
+from .listview import *
+from .radiobutton import *
+from .sliders import *
+from .spins import *
+from .tabs import *
+from .tables import *
+
+from .base import *
+from .callbacks import *
+from .utils import *
+
+try:
+    from .varicons import *
+except ImportError:
+    pass
+
+from ..utils.delegates import *
